Move PaintListener to gfx_traits, Fixes #8834
Adding layers and msg dependency to gfx_traits.
